+++ esp/src/vescController/main/CMakeLists.txt
@@ -1,3 +1,3 @@
-idf_component_register(SRCS "main.cpp" "vescController.cpp" "lidarReader.cpp" "bluetooth_receiver.cpp" "vescLidarUart.c" "rmt-uart/rmt_uart.c" "drive.cpp"
-                       PRIV_REQUIRES driver bt nvs_flash
+idf_component_register(SRCS "main.cpp" "vescController.cpp" "lidarReader.cpp" "wifi_receiver.cpp" "vescLidarUart.c" "rmt-uart/rmt_uart.c" "drive.cpp"
+                       PRIV_REQUIRES driver esp_wifi esp_netif esp_event nvs_flash lwip
                        INCLUDE_DIRS "." "rmt-uart")
