--- esp/src/vescController/main/bluetooth_receiver.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-void init_bluetooth_receiver();
-bool get_manual_control(float &duty, float &steer, bool &emergency);
