+++ esp/src/vescController/main/wifi_receiver.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+void init_wifi_receiver();
+bool get_manual_control(float &duty, float &steer, bool &emergency);
